Move main() method entrypoint to main.dart to follow Flutter standards; allow app to launch from Xcode
